Refactor person to pages app and document #150
